Build enhancements

+++ src/cpic/solver/brute_force_board_solver.cpp
@@ -72,7 +72,7 @@ public:
   BoardState state;
 
 private:
-  ushort nextRow, nextCol;
+  uint8_t nextRow, nextCol;
 };
 }
